Updated the way results are fetched and put into ResultSet objects (see issue #22)

+++ src/main/java/nl/anchormen/sql4es/ESResultSet.java
@@ -124,7 +124,7 @@ public class ESResultSet implements ResultSet {
 		return rows.size();
 	}
 	
-	public void setTotal(int total){
+	public void setTotal(long total){
 		this.total = total;
 	}
